Show (force) reactions onboarding

+++ src/components/StreamersSignin/StreamersSignin.jsx
@@ -15,7 +15,7 @@ import SignInImage from './../../assets/SignIn.png';
 import StreamerDashboardContainer from '../StreamerDashboardContainer/StreamerDashboardContainer';
 import { signInWithTwitch, signUpOrSignInTwitchUser } from '../../services/auth';
 import { getUserToken, subscribeStreamerToTwitchWebhook, subscribeStreamerToMailerLiteGroup } from '../../services/functions';
-import { createStreamerProfile, updateStreamerProfile, userHasPublicProfile } from '../../services/database';
+import { createStreamerProfile, getInteractionsRewardData, updateStreamerProfile, userHasPublicProfile } from '../../services/database';
 import QaplaTerms from '../QaplaTerms/QaplaTerms';
 import { webhookStreamOffline, webhookStreamOnline } from '../../utilities/Constants';
 import { getTwitchUserData } from '../../services/twitch';
@@ -76,17 +76,22 @@ const StreamersSignin = ({ user, title }) => {
             }
         }
         async function redirectUser(uid) {
-            const userHasBeenRedirectedToCreateProfile = localStorage.getItem('userHasBeenRedirectedToCreateProfile');
+            const interactionsRewardData = await getInteractionsRewardData(user.uid);
+            if (interactionsRewardData.exists()) {
+                const userHasBeenRedirectedToCreateProfile = localStorage.getItem('userHasBeenRedirectedToCreateProfile');
 
-            if (userHasBeenRedirectedToCreateProfile) {
-                history.push('/profile');
-            } else {
-                if (await userHasPublicProfile(uid)) {
+                if (userHasBeenRedirectedToCreateProfile) {
                     history.push('/profile');
                 } else {
-                    history.push('/editProfile');
-                    localStorage.setItem('userHasBeenRedirectedToCreateProfile', 'true');
+                    if (await userHasPublicProfile(uid)) {
+                        history.push('/profile');
+                    } else {
+                        history.push('/editProfile');
+                        localStorage.setItem('userHasBeenRedirectedToCreateProfile', 'true');
+                    }
                 }
+            } else {
+                history.push('/onboarding');
             }
         }
